Library Management Project Files Create Successfully

+++ Library_Management_System/src/main/java/com/library/Library_Management_System/Services/UserService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface UserService {
     User saveUser(User user);
     List<User> getAllUsers();
-    User getUserById(Long id);
-    void deleteUser(Long id);
+    User getUserById(int id);
+    void deleteUser(int id);
 }
